pyproject.toml
[tool.poetry]
    authors     = ["Lucas Parzianello <lucaspar@users.noreply.github.com>"]
    description = ""
    name        = "faust-protobuf"
    readme      = "README.md"
    version     = "0.1.0"

    [tool.poetry.dependencies]
        faust-streaming = "^0.10.16"
        grpcio          = "^1.59.2"
        grpcio-tools    = "^1.59.2"
        loguru          = "^0.7.2"
        protobuf        = "^4.25.0"
        python          = ">=3.11,<3.12"
    mypy-protobuf = "^3.5.0"


    [tool.poetry.group.dev.dependencies]
        black = "^23.11.0"
        ruff  = "^0.1.5"

[build-system]
    build-backend = "poetry.core.masonry.api"
    requires      = ["poetry-core"]

[tool.pylint.MASTER]
    # load-plugins = ["pylint_protobuf"]
